Yanh

+++ Laravel/proyectoFinal/app/Http/Controllers/RankingController.php
@@ -7,10 +7,10 @@ use App\Models\Ranking;
 use App\Models\Ranking_analysis;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Hash;
+use Laravel\Sanctum\PersonalAccessToken;
 
 class RankingController extends Controller
 {
-
     public function createRanking(Request $request)
     {
         //funciona correctamente
@@ -33,7 +33,6 @@ class RankingController extends Controller
             "data" => $ranking
         ]);
     }
-
     public function addStudentToRanking(Request $request)
     {
         // Validar los datos recibidos
@@ -66,7 +65,6 @@ class RankingController extends Controller
         // Si el ranking no existe o el código hash no coincide, devolver un error
         return response()->json(['error' => 'No se pudo agregar el estudiante al ranking debido a que el código de la sala no coincide.'], 400);
     }
-
     public function getRankingAnalysesByRankId($id)
     {
 
@@ -79,8 +77,6 @@ class RankingController extends Controller
 
         return $rankingAnalyses;
     }
-
-
     public function getRankingById($id)
     {
         $ranking = DB::table('rankings')
@@ -90,7 +86,6 @@ class RankingController extends Controller
 
         return $ranking;
     }
-
     public function getRankingByTeacher($id)
     {
         $rankings = DB::table('Rankings')
@@ -110,9 +105,6 @@ class RankingController extends Controller
             ], 404);
         }
     }
-
-
-
     public function getRankingByStuden($id)
     {
         //esto funciona perfecto
@@ -140,26 +132,31 @@ class RankingController extends Controller
             ], 404);
         }
     }
-
     public function deleteRanking($id)
     {
-        // Buscar el ranking por ID
-        $ranking = Ranking::find($id);
+        $bearerToken = request()->bearerToken();
+        $user = PersonalAccessToken::findToken($bearerToken)->tokenable;
 
-        if (!$ranking) {
-            // Si no se encuentra el ranking, devolver un error 404
-            return response()->json(['error' => 'No se encontró el ranking especificado.'], 404);
+        if ($user->center) {
+            // Buscar el ranking por ID
+            $ranking = Ranking::find($id);
+
+            if (!$ranking) {
+                // Si no se encuentra el ranking, devolver un error 404
+                return response()->json(['error' => 'No se encontró el ranking especificado.'], 404);
+            }
+
+            // Eliminar todos los registros relacionados en la tabla de análisis de rankings
+            $rankingAnalyses = Ranking_analysis::where('id_rank', $id)->delete();
+
+            // Eliminar el ranking
+            $ranking->delete();
+
+            return response()->json(['message' => 'El ranking y sus registros relacionados han sido eliminados correctamente.']);
+        } else {
+            return response()->json(['message' => 'No estas autorizado.']);
         }
-
-        // Eliminar todos los registros relacionados en la tabla de análisis de rankings
-        $rankingAnalyses = Ranking_analysis::where('id_rank', $id)->delete();
-
-        // Eliminar el ranking
-        $ranking->delete();
-
-        return response()->json(['message' => 'El ranking y sus registros relacionados han sido eliminados correctamente.']);
     }
-
     public function deleteStudentRankingAnalysis($id_rank, $id_student)
     {
         // Buscar el registro que se quiere eliminar
@@ -177,7 +174,6 @@ class RankingController extends Controller
             return response()->json(['error' => 'No se pudo encontrar el registro que se quiere eliminar.'], 404);
         }
     }
-
     public function editRankingPoints(Request $request)
     {
         $id_student = $request->input('id_student');
